Refactored messages module
