+++ trace-viewer/src/messages/cache.rs
@@ -8,8 +8,7 @@ use supermusr_streaming_types::{
     dev2_digitizer_event_v2_generated::DigitizerEventListMessage,
 };
 use tracing::{error, info};
-
-use crate::messages::{CreateFromMessage, DigitiserEventList, DigitiserMetadata, DigitiserTrace};
+use super::digitiser_messages::{FromMessage, DigitiserEventList, DigitiserMetadata, DigitiserTrace};
 
 #[derive(Default)]
 pub(crate) struct Cache {
@@ -18,11 +17,6 @@ pub(crate) struct Cache {
 }
 
 impl Cache {
-    /*pub(crate) fn clear(&mut self) {
-        self.traces.clear();
-        self.events.clear();
-    }*/
-
     pub(crate) fn push_trace(&mut self, msg: &DigitizerAnalogTraceMessage<'_>) {
         info!("New Trace");
         let metadata = DigitiserMetadata {
@@ -40,17 +34,13 @@ impl Cache {
                 error!("Trace already found: {0:?}", occupied_entry.key());
             }
             Entry::Vacant(vacant_entry) => {
-                vacant_entry.insert(DigitiserTrace::create_from_message(msg));
+                vacant_entry.insert(DigitiserTrace::from_message(msg));
             }
         }
     }
 
-    pub(crate) fn iter_traces(&self) -> hash_map::Iter<'_, DigitiserMetadata, DigitiserTrace> {
+    pub(crate) fn iter(&self) -> hash_map::Iter<'_, DigitiserMetadata, DigitiserTrace> {
         self.traces.iter()
-    }
-
-    pub(crate) fn iter_events(&self) -> hash_map::Iter<'_, DigitiserMetadata, DigitiserEventList> {
-        self.events.iter()
     }
 
     pub(crate) fn push_events(&mut self, msg: &DigitizerEventListMessage<'_>) {
@@ -69,7 +59,7 @@ impl Cache {
                 error!("Event list already found: {0:?}", occupied_entry.key());
             }
             Entry::Vacant(vacant_entry) => {
-                vacant_entry.insert(DigitiserEventList::create_from_message(msg));
+                vacant_entry.insert(DigitiserEventList::from_message(msg));
             }
         }
     }
